refactor(codegen/convert): Extract PluginMapper to file
Extracts the implementation of PluginMapper
to its own file in anticipation changes to this code.
